requirements.txt
Django==3.2
ipython==7.22.0
psycopg2==2.8.6
djangorestframework==3.12.4
django-cors-headers==3.7.0
django-filter==2.4.0
Pillow==8.2.0
django-versatileimagefield==2.0
python-magic-bin==0.4.14
sendgrid==6.6.0